added decorators to convert a callable to a command class (#9)
resulting command may be a subclass of `Command`, `Local`, or any desired base

callable will provide the command's functionality, and may be bound as a `staticmethod` (default for `cmd`), as a method of the resulting command class instance (default for `local`), or as a method of the command's immediate parent (default for `cmdmethod` and `localmethod`).

additionally, added a table of contents to the README tutorial, and added documentation of the `entrypoint` decorator.

resolves #6