added Command hooks for the parser's text `formatter_class` and for sub-parser creation
parser characteristics such as the formatter could previously be
controlled by overriding `base_parser`; however, there was no hook for
overriding how sub-parsers were created. this is added as `sub_parser`.

additionally, since a common use-case (especially with regard to the
latter) is to specify a custom text formatter, a hook specifically for
this is provided via class-level attribute, `formatter_class`.